+++ firedantic/_sync/model.py
@@ -5,6 +5,7 @@ import pydantic
 from google.cloud.firestore_v1 import CollectionReference, DocumentReference
 from google.cloud.firestore_v1.base_query import BaseQuery
 
+from firedantic import truncate_collection
 from firedantic.configurations import CONFIGURATIONS
 from firedantic.exceptions import CollectionNotDefined, ModelNotFoundError
 
@@ -50,7 +51,7 @@ class Model(pydantic.BaseModel, ABC):
         self._get_doc_ref().delete()
 
     @classmethod
-    def find(cls: Type[TModel], filter_: dict) -> List[TModel]:
+    def find(cls: Type[TModel], filter_: Optional[dict] = None) -> List[TModel]:
         """Returns a list of models from the database based on a filter.
 
         Example: `Company.find({"company_id": "1234567-8"})`.
@@ -59,6 +60,9 @@ class Model(pydantic.BaseModel, ABC):
         :param filter_: The filter criteria.
         :return: List of found models.
         """
+        if not filter_:
+            filter_ = {}
+
         coll = cls._get_col_ref()
 
         query: Union[BaseQuery, CollectionReference] = coll
@@ -90,7 +94,7 @@ class Model(pydantic.BaseModel, ABC):
             return query.where(field, "==", value)
 
     @classmethod
-    def find_one(cls: Type[TModel], filter_: dict) -> TModel:
+    def find_one(cls: Type[TModel], filter_: Optional[dict] = None) -> TModel:
         """Returns one model from the DB based on a filter.
 
         :param filter_: The filter criteria.
@@ -125,18 +129,10 @@ class Model(pydantic.BaseModel, ABC):
         :param batch_size: Batch size for listing documents.
         :return: Number of removed documents.
         """
-        count = 0
-        col_ref = cls._get_col_ref()
-
-        while True:
-            deleted = 0
-            for doc in col_ref.limit(batch_size).stream():  # type: ignore
-                doc.reference.delete()
-                deleted += 1
-
-            count += deleted
-            if deleted < batch_size:
-                return count
+        return truncate_collection(
+            col_ref=cls._get_col_ref(),
+            batch_size=batch_size,
+        )
 
     @classmethod
     def _get_col_ref(cls) -> CollectionReference: