+++ CHANGELOG.md
@@ -8,6 +8,17 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.2.2] - 2021-04-29
+
+### Added
+
+- Helpers for truncating collections
+
+### Changed
+
+- Make the filter optional for `find` and `find_one`
+
+
 ## [0.2.1] - 2021-03-31
 
 ### Changed
@@ -94,7 +105,8 @@ and this project adheres to
 - Update README.md
 - Update .gitignore
 
-[Unreleased]: https://github.com/digitalliving/firedantic/compare/0.2.1...HEAD
+[Unreleased]: https://github.com/digitalliving/firedantic/compare/0.2.2...HEAD
+[0.2.1]: https://github.com/digitalliving/firedantic/compare/0.2.1...0.2.2
 [0.2.1]: https://github.com/digitalliving/firedantic/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/digitalliving/firedantic/compare/0.1.4...0.2.0
 [0.1.4]: https://github.com/digitalliving/firedantic/compare/0.1.3...0.1.4